add models: Tokens, FavoriteRequests edit: User
