+++ alembic/env.py
@@ -6,7 +6,7 @@ from sqlalchemy.engine import Connection
 from sqlalchemy.ext.asyncio import async_engine_from_config
 
 from alembic import context
-from models.models import Base
+from database.models import Base
 from config_data.config import get_config
 
 # this is the Alembic Config object, which provides
