+++ bot_handlers/save_tokens_command.py
@@ -7,7 +7,7 @@ from bot_base_messages.messages_templates import save_token_mess_templates, err_
 from bot_filters.token_filters import TokenFilter
 from bot_keyboards.keyboards import MakeMarkup, TokenTypeCallbackData
 from bot_states.states import SaveToken
-from models.methods import DBMethods
+from database.methods import DBMethods
 
 database = DBMethods()
