+++ bot_handlers/start_help_commands.py
@@ -5,7 +5,7 @@ from aiogram.fsm.state import default_state
 from aiogram.utils import markdown
 
 from bot_base_messages.messages_templates import info_mess_templates, stickers
-from models.methods import DBMethods
+from database.methods import DBMethods
 
 start_help_router = Router()
 database = DBMethods()
